+++ src/com/twitstreet/market/TransactionMgr.java
@@ -14,7 +14,7 @@ public interface TransactionMgr {
 	
 	public static final int CURRENT_TRANSACTIONS = 10;
 
-	public void recordTransaction(User user, Stock stock, double amount, int operation);
+	public void recordTransaction(User user, Stock stock, int amount, int operation);
 	
 	public List<TransactionRecord> queryTransactionRecord(long userId);
